Add retries to bulk_create_test_applications
So that the frontend looks somewhat realistic in the data its
presenting, we're using a 20-character long identifier when generating
unique emails for the applications created by this task. We're using
SecureRandom.hex(10) for this, which could cause a uniqueness error if
the same email is generated twice for a given database.

Mitigate this by adding a retry mechanism to each query execution,
trying an additional 3 times before raising an error.