Enable course options to be held open
In order to support providers who want to test live courses on the
sandbox while we're between cycles, allow course options to have a
hold_open property. Options with this property will always appear to
have vacancies.

- create the boolean field on the model
- redefine the enum-created `no_vacancies`? method to respect it
- redefine the `available` and `vacancies` scopes to respect it

The spec shows that it's now possible to apply to a course that's full
as long as it has the hold_open option set.