Enable course options to be held open
In order to support providers who want to test live courses on the
sandbox while we're between cycles, create a feature flag that
forces all courses to appear to have vacancies.

Note that the :vacancies scope needs to be defined after the enum,
because the enum creates its own scope we need to override.

The spec shows that it's now possible to apply to a course that's full
as long as the flag is flipped.

Show a notice on the provider and course selection screens in the
candidate UI indicating that this feature is turned on.