Use hobby-basic for Heroku review apps dbs
This will raise the row limit to 10 million, which will eliminate the
row limit warnings we've been receiving.

We've previously tried doing the same in
https://github.com/DFE-Digital/apply-for-teacher-training/pull/1817.

That didn't work, because Heroku has an ephemeral restriction to
disallow certain add-ons on review apps. After contacting support,
they've lifted the restriction for the `dfe-bat` team.