Tidy up error handling in DeferOffer
The error handling here is not consistent.

- We return `false` and populate the `errors` if the state transition
is illegal. This is something that shouldn't happen in the normal user
flow, so it's legitimate to raise an error that will notify us.
Currently the provider user would see a message "The application is not
ready for that action", which isn't helpful
- We're using `update` to set `status_before_deferral` and
`offer_deferred_at`, which will return false if the validation fails,
which we ignore - because we just return true in the last line of the
method. This makes sure that we fail if the update fails.
- Rename the `save` method to indicate that it'll raise an error if it
fails.
- Remove the error checking in the controller because it will never be
called anyway.