Fix N+1 queries in provider users form
We currently make 1400 queries to build the new / edit provider user
page, which is a looot. And that's with 350 providers! If we sync all
2000 providers, this would increase to 7,811.

This fixes the N+1's by:

- Memoizing the possible_permissions, because we're calling this method
twice.
- Fetching all the current user's permissions in one go, and doing a
lookup in Ruby
- Populating the `provider` object in case we haven't found a
`ProviderPermissions` record, so that we don't have to make a query for
that again.

This reduces the number of queries to 3 or 4 or something.