Find or create candidates in process matching data spec
We were intermittently getting the following error:

ActiveRecord::RecordNotUnique:
  PG::UniqueViolation: ERROR:  duplicate key value violates
  unique constraint "candidates_pkey"
  DETAIL:  Key (id)=(44) already exists.

In this test we are creating Candidates with specific IDs, which
have to correspond to the data in matching_data_example.csv
Occasionally those IDs were already used causing the error.

To fix this we find the first user with a particular id (we don't care
about email address) but if there isn't one we need to create it
and specify email_address as we validate it's presence.