Add service object for sending decoupled references emails
Currently, if a candidate navigates back to a stale page that asks them
to confirm if they want to send a reference request email, and they
click 'yes', we send an email each time. We want to put a check in place
to ensure only one email may be sent per referee.

Introduce a separate service object to accommodate the change in behaviour
from the existing SendReferenceRequest.

Differences in this implementation:

- Pass the flash object in to conditionally set success/failure messages
- Check the state of the reference and then either send the request or
  set a failure message.