Update controllers, components and views and Add test for cancel
As part of the decoupled references feature candidates require the ability to cancel references. This fix extracts the logic from the existing referees feature and repurposes it to work within the new feature. Cancel and confirm cancel actions are added to the review controller along with modifications to content accordingly.

Remove redundant code from controller

Remove merge markers