Remove defunct CandidateMailer actions
As a result of the Decoupled references feature the function of references has been redesigned. As such redundancy, needs to be removed from the codebase. Several actions have been removed from the model and the files they touch have
also been amended or removed if they are affected.

Remove further tests