Show validation error to candidates for references with same email
When a candidate submits a reference with an email that is the same as
that of the candidate, we should show a validation error.

This adds the validation error to the form model, which is copied from
the ApplicationReference model. The validation error is also duplicated
in the localisation file.

I also changed `:email_address_unique?` to not use a `?` as it doesn't
technically return a boolean, and I fixed up another confusing variable
name.