Add invariant check for weird reference state
An important property of the system is that submitted application forms
not have any outstanding reference requests. Such a state could lead to
unexpected reference chase emails being sent, or a bad user experience
if a referee tries to provide a reference when it's no longer needed.

This invariant check looks for this state in the database and sends a
message to Sentry with the ids of any applications it finds.