Show "Invalid data" when a UCAS match is invalid
If we encounter a UCAS match with an invalid course code, provider code
or a course that doesn't exist in the current recruitment cycle, we
currently raise an error
(https://sentry.io/organizations/dfe-bat/issues/1937345465).

This changes the flow so that we check if the match is valid, and
output an error status instead of crashing.