Export only application in the current recruitment cycle
During the October 2020 Data Matching DSA Review we spoke with UCAS about two
ways handling data in different recruitment cycles. The first option was to
only send through candidates from the 2021 cycle going forward. The second was
adding a 'recruitment cycle year' field in the data we send over and receive,
which would require an amendment to the DSA. As discussed, both parties
favoured the former.

For this reason we need to change UCAS export to only send candidate data from
the current recruitment cycle. 