Improve search for application forms
Our current "ILIKE" based search isn't very good, because it doesn't do
fuzzy-like matching. For example, if the user is named "Sam Norman
Seaborn", a search for "Sam Seaborn" will not return anything.

Luckily, Postgres is a beast and supports very cool full-text search
out of the box. To make things even easier we can use the `pg_search`
gem, which allows us to create a scope for nice searching.