Specify node and yarn versions
Adding these to the `package.json` file makes it clearer which versions of Node.js and Yarn are required.  `yarn install` will throw an error if incompatible versions are installed.

The `^` prefix specifies a [caret range](https://docs.npmjs.com/cli/v6/using-npm/semver#caret-ranges-123-025-004) – which means that higher minor or patch level versions are allowed, but not a higher major version (as this is likely to be backwards-incompatible), following the [semantic versioning](https://semver.org) convention.