TestApplications cannot create timestamps in the future
Because of the way that we moved @time around inside TestApplications,
it was possible for us to create timestamps in the future.

The total number of days that "fast_forward" caused us to move could
bring the "current" time ahead of the real current time.

This caused issues with test data for vendors.

Instead of passing a range to fast_forward, let it decide how far
forward to go based on how long there is between "now" and the last
moment the application could have been edited. Hopefully this means we
create a realistic timeline that conforms to the Cycle Timetable.