Make tests for realistic data less strict
With the changes to the way we handle time in the generation of test data, it's no longer sensible to ask that events are evenly spaced. This changes the requirement to be that things happen in a realistic order. Also allow the time to jump in fast_forward to be lower than 2