Revise the international degree flow
Improve the grade step of the flow for international degrees by framing
the question differently and changing the order of the radio buttons.

Instead of asking 'What grade is your degree?' we now ask them 'Did your
degree give a grade?' and relabel the options accordingly. The most
useful option (which allows the candidate to enter a grade) is moved to
the top.

Notable changes:

- For consistency, we map the new radio button values to the old values
  that we were saving in the database (see `DegreeGradeForm`).
  Specifically -

 |-----------------------------------------|
 | Submitted form value | Persisted value  |
 |--------------------- | -----------------|
 | 'No'                 | 'Not applicable' |
 | 'I do not know'      | 'Unknown'        |
 |-----------------------------------------|

- For clarity, the `DegreeGradeForm#fill_form_values` method has been
  refactored to have two different approaches for UK and international
  degrees.
- The page title on the grade step dynamically changes based on whether
  or not the degree is complete. This gives us four possible page titles
  based on permutations of UK/International and Predicted/Complete.