Remove international address Feature flag and refactor specs
The international Feature flags are now blocking progress and need to be removed. I have removed the international feature flag from all mentioned places in the code base. I have had to create an additional method in the candidate helper and update the current contact details to work independently within the system tests.