Fail gracefully when candidate uses back button
After creating a new other qualification (or updating an existing one) a
candidate can use the back button to return to the details form and
attempt to re-submit. This fails because we don't have the qualification
type value in the wizard state and so validation errors appear for
fields that are not on the details form.

We can't easily make the back button work as intuitively as it did for
the save-as-you-go model so this isn't an ideal solution - but we can
indicate to the candidate the 'correct' way to update a qualification.