Flag "Action needed" immediately after requesting removal from UCAS
The flow was reviewed by the Transition Manager/UCAS engagement lead who
specified that UCAS should enact the removal immediately. Therefore we want to
flag a match as "Action needed" as soon as support agent clicks the button that
they requested withdrawal from UCAS.

The instructions for confirming removal from UCAS say: "We need to ensure that
UCAS have removed the candidate's duplicate application from UTT by the end of
the day. Continue to contact UCAS until this has been confirmed." to reflect
that.