Move without_slack_message_sending into StateChangeNotifier
This makes it a bit clearer where the message sending actually
happens, and lets us dedupe some code we're about to reuse.

Make the language here more consistent: remove mention of slack from the
RequestStore key and talk only about notifications, not messages, in
line with the name of this class.