Cancel references for cancelled applications
The `DetectInvariants` checker checks if there are requested references
that belong to application choices that haven't been submitted yet.
This is impossible because you can't submit applications without all
references being either cancelled, declined or with feedback.

However, in QA/Sandbox we also have application choices in the
cancelled state. This state has been removed, but it still exists on
prod and so we need it as test data too. This is currently raising an
error:

https://sentry.io/organizations/dfe-bat/issues/1990257169