Add UCAS holidays to calendar
UCAS have made us aware of their breaks in winter 2020 and spring 2021. Add these to the calendar to ensure RBD dates in apply are fair.
Also fix RBD calculation tests - we need to calculate the rules for time limits when the method is called, rather than when the module is loaded