Refactor determining the section
This doesn't rely on the URLs, which can change. It also doesn't let
users change the page title by fiddling with the URL - if this happens
or if we don't know the section, we default to "this section".

This also fixes a bug where the section would disappear when there are
validation errors. And simplifies the passing through of `path` and
`page_title`.