Fix Bug in carried over application Application Review
When a user completes the course choices section after carrying over an application there is a bug which prevents
user from adding more than one course choice if the section is comopleted. This is due to the incorrect component
being rendered on the review screen. Changes in this PR address this by refactoring the show view to show the
correct component if the user has a previous application. Specs updated to mirror situation.