Update geocoding check in ApplicationForm
Also check if `address_type` has changed when deciding if geocoding
should take place. In practice, the system changes `address_type` if
`country` gets updated. The ApplicationForm model has no knowledge of
this fact, however, so we just check for changes on the field
independently.

This has the side effect of making one of the tests clearer and a better
representation of the actual behaviour being tested.