Send initial/reminder UCAS match emails when processing matching data
Decided to send emails after processing matches instead of doing it in the
clock because if there is a problem with our sync we don’t want to be chasing
candidates in case they have taken an action.

It will be still useful for see how many matches required an action, hence
sending emails after the stack notification.