Format API grade example as a string, not an array
Some parts of OpenAPI do support multiple examples under an "examples:"
key but properties unfortunately don’t. The old code rendered an array
in the docs which implied that clients should send an array.

Put the extra information in the description and leave a valid example
in the example field.