Use new method of dealing with tokens for candidates
This makes the candidate magic link authentication use the new
AuthenticationToken class, which means that if they request a sign-in
email multiple times they will be able to use all of the links they've
been sent (within an hour).

This prevents a cycle where the candidate requests a sign in email,
then waits, then requests one again, receives the first email, clicks
on the link, finds it expired, requests a third email, then clicks on
the second email link, etc ad infinitum.