Use StripWhitespace wherever we currently strip params
There are several controllers in the CandidateInterface where we are
already stripping spaces by transforming the values of params hashes
just prior to passing them on to form objects.

In all of these cases, replace the call to `transform_values` with a
call to StripWhitespace. Do this via a convenience method
`strip_whitespace` added to the CandidateInterfaceController.