Remove the withdraw link froms states with accepted offers
At the moment, a provider is presented with a link to withdraw
offers for application choice in the deferred and pending conditions
states. When they try and confirm the withdrawalthey are presented with
an error 'The application is not ready for that action' as we do not
allow that transition of state to occur.