Fix specs broken by updating the ApplicationResponseCache
Now that saving an ApplicationChoice runs
VendorAPI::SingleApplicationPresenter, we have to take care not to save
inconsistent data in tests, as this can break the presenter.

Whitelist the association with ApplicationResponseCache or we end up
having to include() it everywhere, which is too noisy - accept the
(presumably) very slight performance hit instead.