Only show status change events for statuses visible to providers
Fixes:
```
ActionView::Template::Error undefined method `match' for nil:NilClass
```
in ProviderInterface::ApplicationTimelineComponent
https://sentry.io/organizations/dfe-bat/issues/2132714011/events/5d13bf5b27cd4c9b9ec9760efcc9230e/?project=1765973

The `title_for(status)` was nil for statuses not visible to providers because
they are not defined in the TITLES constant.  Those type of events (e.g. when
status changes from "awaiting_references" to "application_complete") should not
be included in the timeline.