Set current Redis in an initializer
Currently we interact with Redis in a few different ways throughout the
codebase:

- Sidekiq by default uses the value of REDIS_URL to configure its server
  and client connections
- We have direct invocations of `Redis.new`, which returns a connection
  pointing at whatever REDIS_URL is set to
- We also directly invoke `Redis.current`, which does the same

Standardise all of these as follows:

- Set `Redis.current` in an initializer, using the URL returned by
  ApplyRedisConnection
- Use `Redis.current` consistently wherever we require access to the
  Redis connection, including when configuring Sidekiq

Not only does this give us a slightly more consistent interface, it also
ensures that the correct Redis instance is being used in environments
where REDIS_URL is not set (ie - wherever we want to do parallel test
runs).

An example of why this is required - if REDIS_URL isn't set, `Redis.new`
will default to database 0. As per the implementation of
ApplyRedisConnection, we want more fine-grained control over which
database is used when running tests.