Add a Redis connection wrapper to support parallel test runs
In order to support parallel tests, we need a different Redis instance
for each concurrent process. Currently, we declare the Redis location
for a running instance of the app via the REDIS_URL environmental
variable. We need a slightly more flexible mechanism than this to set a
different Redis database per process.

This commit adds a class that returns the appropriate Redis URL
depending on the environment. It makes checks in the following order and
does one of these things:

- Use REDIS_URL if present. This means that existing environments that
  have this set will continue to use the same connections.
- Look for TEST_ENV_NUMBER, which is set by the parallel_tests gem.
  This number will be unique in each process, and is used to pick a
  numbered Redis database. By default, Redis has databases 0-15
  available, and this code will choose a database from 1 onwards.
- Check if this is a single-process test run, in which case use the
  arbitrarily-chosen Redis database 9.
- Fall back to using Redis database 0 if none of the above apply (this
  may be the case in the development environment, for example).