Add received at columns to Application Reference Data Extract
As a requirement on the current Application Reference Data Extract, business analysts require
the capacity to view the time at which references have been completed. This solution builds on the
existing data extract model and adds the columns along with a method to find find the received_at
time in the audit log.

rubocop