Use let_it_be/before_all from test-prof gem to speed up spec
These methods hook into RSpec's before(:context) to allow database
records to be set up once before executing examples in the group. For
specs which create a lot of test data in their setup, this significantly
reduces the amount of time spent in the database. Any modifications to
these records are automatically rolled back between examples, and an
option is available to automatically reload the ActiveRecord object as
well.

Use these methods in ucas_matched_application_spec. This change reduces
the runtime from 6.72 seconds to 0.5 seconds (tested on a machine with
AMD Ryzen 7 3700X).