Avoid inadvertent rejection of applications
We've got an application that was inadvertently rejected by a provider.
I think it's some kind of back-button issue.

Normally, this wouldn't be a problem, because our state machine
protects against weird transitions. However, the state transition to
reject an offer and the transition to withdraw an existing offer is the
same (#reject!). This means that the `RejectApplication` service was
invoked, and not the `WithdrawOffer` service. Because of this we don't
set the `offer_withdrawn_at` attribute, which in turn crashes the offer
page.

This changes the transition so any inadvertent invocation of the
`RejectApplication` service for offered applications will result in a
crash rather than an inconsistent state. Hopefully it will allow us to
find the scenario where this occurred so we can fix it.