Add received_at column to references table
As part of technical debt we use several database queries throughout our export files
to call the value for the time a reference moves into feedback provided status from the audit
log. Rather than do this each time it makes sense to have a separate column in the database.
This PR adds this column and it's migration.