Configure Redis for parallel test execution
A prior commit dealing with this was reverted as part of an
investigation into performance issues in production. Although seemingly
unrelated to the issue, it did prompt a review of how we're configuring
Redis for parallel tests.

This commit once again adds support for each parallel process to use a
different Redis database. However, it takes a different approach -

- Rather than configuring connections directly, we continue to rely on
  the REDIS_URL set in a given environment.
- Use ClimateControl to stub the value of REDIS_URL when running tests
  in parallel, ensuring we have uniqueness without affecting the
  configuration of anything else in the codebase.

In addition, this commit updates .env.example and .env.development to
encourage the convention of using Redis database 0 in development,
leaving databases 1 through n available for the test environment.