Copy spec for R4R dashboard from Feature Metrics Dashboard
Our team requires the ability to view the metric for the reasons for rejection across the application.
As such this feature is to create a dashboard to measure these reasons via the support UI. This PR
creates the first page of this dashboard.

This commit begins the process by copying and customising the system spec from the adjacent Feature
Metrics Dashboard.