Update incorrect logic based on PR feedback
Ensure that candidates are sent ot the review page if they have
already added a job or have given an explanation as to why
they have no work history & link to the start page if not.

Also fixes the incorrect logic on rendering a 404 if the
flag is not swtiched on. Previously it 404'ed if it
was inactive.