Update the new bool to false when a candidate uses the old flow
Prior to the flag being switched on, exisitng candidates wil still
see the old flow. This commit ensures that as soon as a candidate is
presented with the first page of the existing flow we mark them
as having seen it, ensuring that they will not see the new flow
once it is fully implemented and the flag is switched on.