Add Percentage value to Dashboard
After discussion with stakeholders it was clear that before_this_month wqs not useful and that instead
a percentage of total reasons would be more helpful. This commit adds logic to the view component to
calculate this value and then rejigs the front end based on styling comments. The spec has also been updated to reflect the changes to the front end, to include a test for the percentage value.