Add Proper Pagination to SyncAllProvidersAndCourses
Pagination is currently implemented on this TeacherTRaining API service in a sub optimal way.
Currently, the call method on this object use a begin rescue block which counts pages from 1 until
it hits an error and stops. This PR refactors the code to paginate to the correct number of pages
exactly. A unit test has also been added along with some stubbed JSON pages to test this functionality