Fix issue with grades when changing BTEC to GCSE
After creating a BTEC other qualification and then changing the
qualification type to GCSE the details form still shows the grade
options for BTEC (radio buttons). This is a simple change to ensure that
these buttons are only shown when the qualification_type is 'Other' and
the other_uk_qualification_type is 'BTEC'.