Use the total structured rejection count to calculate the % shown
Previously, we were calculating the total number of reasons for rejection
then doing reason/total_number_of_rejection_reasons_given * 100.

This doesn't really work for a couple of reasons. Firstly,
while is indicates rejection reasons are more common, this can
be ascerntained by just looking at the total number of rejection
for each section.

Secondly, the interested in future applications and other advice or
feedback sections will distort the percentages for the actual rejection
reasons.

These two sections are not actually rejection reasons, so if we were to
use the previous implementation, should not be included.