Added restructured work experience component
In order to use the new work history design a new component has been built to display the individual work experience entries on the work history review page, in place of the existing summary card component. This covers the role, organisation, time period and if the job had relevant skills.