Don't show Decline by default at if it hasn't been set yet
The component was breaking when 1 offer was made but the candidate was still
awaiting decisions from other providers as the DBD was nil.

It is possible to have an application choice is offered state without decline
by default date, as DBD is only set it once all the candidate's application
choices have either an offer, been rejected by provider, withdrawn by the
candidate, or have the offer withdrawn.

The `decline_by_default_at` for all offered choices is set when the last
offer is made.