Mix state-mixing bug in "Other qualifications" section
The "Other qualifications" section uses our wizardry functionality.
This means that we have multiple pages with forms, and save the
candidate input into an "intermediate data service" first. Once the
candidate is finished with all the pages we persist the data to the
database. This allows us to ask the user for less information per page
but avoid invalid and inconsistent data in the database.

The initial implementation of this feature saved the data provided by
the user into a singular Redis field
(`candidate_user_other_qualification_flow-#{current_candidate.id}`).
This means that every time we load a qualification form page, we have
to clear the state to make sure that if we visit a different form
page the wrong data isn't displayed. This is very hard to do, and we
currently are seeing the following bug:

- Create a application with 2 "Other Qualifications"
- Click on "Change" for the first qualification
- Use the back button in your browser
- Click on "Change" for the second qualification
- The form is erroneously populated with the data for the first
qualification, not the second

This PR changes the approach, to avoid the state clearing in most cases
(this is proof of the lesson is that cache invalidation is hard so best
avoided). By using key per record, it's not possible to share data
between the forms.

The only place where we need to explicitly clear the state is when the
candidate visits the "Add other qualification" page (the one that lets
you choose the type). Without the reset, the second time a candidate
clicks the "Add" button they'll see the data they've entered the first
time. I've considered avoiding this by generating an ID every time you
visit the page, persisting this in the session, and using this as the
key suffix instead of "new". I suspect that's more trouble than it's
worth though.