Ignore query params in the tab navigation component
This lets us set default query options for filters etc under each tab
and keep the tab highlighted when those options change.

What if the tab title doesn't apply when the query parameters change,
though?  I don’t think this is a problem because if query values would
cause the content to be in a different tab - ie under a different
heading altogether - then that ought to be a different URL anyway.