Fix vulnerability in fallback support authentication
Recently we've started using the `AuthenticationToken` model for
support fallback authentication as well.

The method to look up the token doesn't scope the lookup by token to
support users. This means that an attacker could request a sign in link
on the candidate interface, and craft a URL to use the token in the
link to gain access to the support console.