Updated work history break form tests
The tests for the work history break form were only checking
the previous candidate flow. The tests have now been moved over to
cover the work history break form, on the new restructured work history
flow.

In addition, two of the tests in the gap component spec have been updated
to also check that the correct path is rendered by the links.