Application review shows restructured work history
The work history section on the candidate interface has been updated
with a new design and flow. This design will now also appear on the
application review page. Parts of the CSS have been moved around in
order for the section to format correctly on both the application
review page and the work history review page.