Added a flash message for unexplained work history breaks
As the application review page only checks if the work history
section has been completed or not, validation was required to ensure
that all gaps have either been explained with a break or that a job
has been added to cover that period of time. A flash message has been
added to the work history review page that will appear if the candidate
tries to mark the section as completed with unexplained gaps.