Restore interviews in the provider activity log using joins
Our last attempt at including interviews produced SQL which specified
the OR logic needed for an .own_and_associated_audits query in the
WHERE clause of a query and included the application_choices visibility
rules twice. By making this OR logic a part of the INNER JOIN between
audits and the application_choices from GetApplicationChoicesForProviders,
we can scan the audits table more efficiently and take advantage of the
existing [auditable_type, auditable_id] and [associated_type, associated_id]
indexes.

Besides speed, this approach should work well for any other
ApplicationChoice-associated audits we add in the future, similar to
interviews. We filter so much on ApplicationChoice statuses and other
fields because not all state transitions are relevant for Manage, but
associated models (such as interviews) should either be included in the
feed for all their changes or not at all.