Add new notification banner component
- don’t display govuk-body if neither a link or a secondary message are present
- set role depending on notification type as specified in the GDS design system
- supports html in both primary and secondary message

Role explanation:

The distinction of using role `alert` or `region`

alert: `The alert role can be used to tell the user an element has been dynamically updated. Screen readers will instantly start reading out the updated content when the role is added.`

region: `The region landmark role is used to identify an area in the document that the author has identified as significant. It is used to provide a generic landmark for people to be able to navigate to easily when none of the other landmark roles are appropriate.`