Added the ExportExport class
So that we can keep track of what is in each export, we are building
an export that contain the name of each export and what column names
are included.

The ExportExport class has been created. This function of this class
is to collect the class name of each export within the DataExport
class, then use this to store each column name that is populated by
the data_for_export method. This will then be stored to a hash that
will then populate the CSV file that is returned to the user.