WIP: ExportExport class now returns the export name
The ExportExport class will now exclude itself when generating the
data of each export available. An argument has also been added
to the data_for_export method - this will be used as a flag to break
out of the loop if we only want to return the first row.

To test this is working without having to run all the exports, the
ApplicationsExport, ActiveProviderUsersExport & CandidateApplicationFeedbackExport
have been modified to include this argument and the rest of the exports
have been commented out to avoid them being run when the ExportExport
is used.