WIP: Updated the individual export classes
In order to use the functionality of the ExportExport class each
individual export requires a data_for_export method, which takes
a single argument that's used as a flag to determine if the loop
should only run once.

The individual export classes have been updated to ensure they have
a data_for_export method. Some of these classes had it as an alias,
but were unable to accept an argument. A break has also been added
to the method.