Address PR comments
This commit adds a couple of tests. One to ensure presence is checked
on the work histoy breaks start date and one to ensure we mark
the section is incomplete when a work break is updated.

It also fixes an issue with checkboxes not having the value
retained in the job create action.