Exclude the application status transition from cancelling an interview
When an interview is cancelled the application reverts to awaiting_provider_decision
and this update appears in the timeline and activity log. We should filter this as the
audit entry for the cancelled interview is more accurate in describing what has happened.
This is also consistent with filtering the application status change when creating an interview.