Handle non-integer values for app choice id filter
This was causing a database error whenever `application_choice_id` was
not a numeric string, e.g. 'ABC' or 'abc123'. There is no validation in
the filter so the correct thing to do is to treat this as a non-matching
value by coercing it to an integer before it gets to the database.
Because `'ABC'.to_i` evaluates to `0` it leads to no matching
application choices which is the correct result I think.