Rename the rejection columns in the E&D export
As part of the work to understand and improve the data currently
available for export it was identified that the rejection reason
columns could potentially lead to confusion in the E&D export.

The naming has now been updated to clarify that each column
relates to a different application choice.